rename rangefilter file
